Fix : Procedure.Progress被截断的问题

+++ Icy/Packages/fun.program4.icy.base/Runtime/Procedure/Procedure.cs
@@ -61,7 +61,7 @@ namespace Icy.Base
 		/// <summary>
 		/// 获取归一化的执行进度，跳转Step也包括
 		/// </summary>
-		public float Progress => (_CurrStepIdx + 1) / _Steps.Count;
+		public float Progress => (float)(_CurrStepIdx + 1) / _Steps.Count;
 		/// <summary>
 		/// 当前Step
 		/// </summary>
